Moved plotting functions to plotting.py
